# docs/api/vedal-admin-openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: sales@vedal-med.ru
    name: VEDAL
    url: tel:88006003449
  description: |
    Контракт сайта vedal-med.ru с серверной частью портала.

    Двери и их правила:

    - `/api/public/v1/**` — чтение опубликованного. Ответы кэшируются
      на пять минут (`Cache-Control: max-age=300, public`); файлы
      документов не кэшируются вовсе, потому что снятая с публикации
      редакция не должна жить в кэшах прокси.
    - `/api/forms/v1/**` — приём заявок. Отвечает `202 Accepted`:
      заявка принята, дальше её разбирает менеджер.
    - `/api/assistant/v1/**` — вопрос ассистенту.

    Ошибки во всех дверях — `application/problem+json` по RFC 9457.
    Разбор по полям формы приезжает в расширении `fields`.

    Кросс-доменные запросы разрешены только источникам из
    `vedal.web.allowed-origins`, без передачи учётных данных: у
    публичных дверей нет ни cookie, ни сессии. Повторную отправку
    заявки различает заголовок `Idempotency-Key`, а не пользователь.

    Лимит частоты считается по адресу клиента, у каждой двери свой
    бюджет: формы — 5 обращений за 10 минут, ассистент — 20.
    Превышение отдаёт `429`.
  title: VEDAL Portal — публичное API
  version: v1
servers:
- description: Generated server url
  url: http://localhost:8081
...
paths:
...
  /api/admin/v1/clients/{id}/history:
    get:
      operationId: history_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminInteraction"
          description: OK
      security:
      - keycloak: []
      summary: История переписки и звонков по клиенту
      tags:
      - "Админка: клиенты"
    post:
      description: "История только дописывается: правки и удаления здесь нет. Исто\
        рию, которую можно поправить задним числом, незачем вести."
      operationId: addToHistory_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNewInteraction"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInteraction"
          description: Created
      security:
      - keycloak: []
      summary: Дописать в историю клиента
      tags:
      - "Админка: клиенты"
...
  /api/admin/v1/deals:
    get:
      operationId: deals
      parameters:
      - description: "Фильтр по воронке: sales, dealer, service. Пусто — все."
        in: query
        name: pipeline
        required: false
        schema:
          type: string
      - description: Фильтр по стадии. Стадия из чужой воронки — «409».
        in: query
        name: stage
        required: false
        schema:
          type: string
      - description: "Все сделки одного клиента. Сужает наравне с воронкой, а не в\
          место неё."
        in: query
        name: clientId
        required: false
        schema:
          type: string
          format: uuid
      - description: Логин ответственного; «-» — без ответственного.
        in: query
        name: owner
        required: false
        schema:
          type: string
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - description: "Размер страницы, не больше 200."
        in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
          description: OK
      security:
      - keycloak: []
      summary: "Сделки постранично, свежие сверху"
      tags:
      - "Админка: сделки"
    post:
      description: "Сделка из заявки заводится другой дверью — `POST /leads/{id}/convert`."
      operationId: create_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNewDeal"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: Created
      security:
      - keycloak: []
      summary: Завести сделку по существующему клиенту
      tags:
      - "Админка: сделки"
  /api/admin/v1/deals/pipelines:
    get:
      description: "Набор стадий у каждой воронки свой. Форма рисует выбор из них\
        , а стадию из чужой воронки не примет ни домен, ни схема."
      operationId: pipelines
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineView"
          description: OK
      security:
      - keycloak: []
      summary: Воронки и их стадии
      tags:
      - "Админка: сделки"
  /api/admin/v1/deals/{id}:
    get:
      operationId: deal
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: OK
      security:
      - keycloak: []
      summary: Карточка сделки
      tags:
      - "Админка: сделки"
    put:
      description: "Стадия меняется отдельной дверью: перевод по воронке — это соб\
        ытие, а не правка поля."
      operationId: update_4
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDealForm"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: OK
      security:
      - keycloak: []
      summary: Править карточку сделки
      tags:
      - "Админка: сделки"
  /api/admin/v1/deals/{id}/attachments:
    post:
      description: "Прикладывается ссылка на карточку документа, а не копия файла\
        . Несогласованный документ приложить нельзя — `409` с разбором причины."
      operationId: attach
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDealAttach"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: OK
      security:
      - keycloak: []
      summary: Приложить согласованный документ к сделке
      tags:
      - "Админка: сделки"
  /api/admin/v1/deals/{id}/attachments/{documentId}:
    delete:
      operationId: detach
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: documentId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: OK
      security:
      - keycloak: []
      summary: Отцепить документ от сделки
      tags:
      - "Админка: сделки"
  /api/admin/v1/deals/{id}/history:
    get:
      operationId: history_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminInteraction"
          description: OK
      security:
      - keycloak: []
      summary: История переписки и звонков по сделке
      tags:
      - "Админка: сделки"
    post:
      description: "История только дописывается: правки и удаления здесь нет."
      operationId: addToHistory_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNewInteraction"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInteraction"
          description: Created
      security:
      - keycloak: []
      summary: Дописать в историю сделки
      tags:
      - "Админка: сделки"
  /api/admin/v1/deals/{id}/quotes:
    get:
      operationId: quotes_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminQuoteRow"
          description: OK
      security:
      - keycloak: []
      summary: "КП по сделке, свежие сверху"
      tags:
      - "Админка: сделки"
  /api/admin/v1/deals/{id}/stage:
    post:
      description: |
        Стадия обязана быть из воронки этой сделки. Перевод в проигранную стадию
        требует причины: воронка без причин показывает, сколько потеряли,
        и молчит о том, почему.
      operationId: moveTo
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDealStage"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: OK
      security:
      - keycloak: []
      summary: Перевести сделку на другую стадию
      tags:
      - "Админка: сделки"
...
  /api/admin/v1/leads/{id}/convert:
    post:
      description: |
        Заводит сделку в выбранной воронке. Клиент либо указывается существующий,
        либо заводится из данных заявки — портал не ищет совпадения сам: слить
        две карточки потом можно, разделить ошибочно слитые уже нет.

        Заявка разбирается один раз. Повтор — `409`: две сделки по одному
        обращению обе попали бы в аналитику.
      operationId: convert
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLeadConversion"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminDeal"
          description: Created
      security:
      - keycloak: []
      summary: Разобрать заявку в сделку
      tags:
      - "Админка: заявки"
  /api/admin/v1/leads/{id}/history:
    get:
      operationId: history
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminInteraction"
          description: OK
      security:
      - keycloak: []
      summary: История переписки и звонков по заявке
      tags:
      - "Админка: заявки"
    post:
      description: "История только дописывается: правки и удаления здесь нет."
      operationId: addToHistory
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNewInteraction"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInteraction"
          description: Created
      security:
      - keycloak: []
      summary: Дописать в историю заявки
      tags:
      - "Админка: заявки"
...
components:
  schemas:
...
    AdminDeal:
      type: object
      description: Карточка сделки вместе с вложениями.
      properties:
        amount:
          type: number
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AdminDealAttachment"
        clientId:
          type: string
          format: uuid
        clientName:
          type: string
        closedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        currency:
          type: string
        id:
          type: string
          format: uuid
        leadId:
          type:
          - string
          - "null"
          format: uuid
          description: "Заявка, из которой заведена сделка. Пусто у сделки, заведё\
            нной руками."
        lostReason:
          type: string
        lostStages:
          type: array
          description: "Стадии, которыми эта воронка заканчивается отказом. Перево\
            д в такую стадию требует причины проигрыша: форма спрашивает её до наж\
            атия, а не после отказа."
          items:
            type: string
        owner:
          type: string
        pipeline:
          type: string
        productSlug:
          type: string
        stage:
          type: string
        stages:
          type: array
          description: "Стадии этой воронки, по порядку. Форма рисует выбор из них\
            , а не свободное поле."
          items:
            type: string
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          description: Версия карточки. Её надо вернуть в форме правки.
        wonStages:
          type: array
          description: "Стадии, которыми эта воронка заканчивается успехом."
          items:
            type: string
    AdminDealAttach:
      type: object
      description: Приложить согласованный документ к сделке.
      properties:
        documentId:
          type: string
          format: uuid
      required:
      - documentId
    AdminDealAttachment:
      type: object
      description: |
        Приложенный документ. Хранится ссылка на карточку, а не копия файла:
        копия разошлась бы с оригиналом при замене ревизии.
      properties:
        attachedAt:
          type: string
          format: date-time
        attachedBy:
          type: string
        documentId:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
    AdminDealForm:
      type: object
      description: Правка карточки сделки. Стадия меняется отдельно.
      properties:
        amount:
          type: number
          example: 2650000.0
          minimum: 0
        currency:
          type: string
          pattern: "^$|^[A-Z]{3}$"
        owner:
          type: string
          maxLength: 200
          minLength: 0
        productSlug:
          type: string
        title:
          type: string
          example: Поставка двух систем VEDAL R2
          maxLength: 300
          minLength: 0
        version:
          type:
          - integer
          - "null"
          format: int64
          description: Версия прочитанной карточки.
      required:
      - title
    AdminDealStage:
      type: object
      description: |
        Перевод сделки на другую стадию. Воронка не меняется: сделка,
        переехавшая из продаж в сервис, — это две разные сделки.
      properties:
        lostReason:
          type:
          - string
          - "null"
          description: "Причина проигрыша. Обязательна при переводе в проигранную\
            \ стадию: воронка без причин показывает, сколько потеряли, и молчит о\
            \ том, почему."
          maxLength: 500
          minLength: 0
        stage:
          type: string
          description: Стадия из воронки этой сделки. Список приходит в карточке и
            в `GET /deals/pipelines`.
          example: qualified
          minLength: 1
      required:
      - stage
...
    AdminInteraction:
      type: object
      description: |
        Запись истории. Текст — персональные данные: наружу и в топики
        не уходит, туда идёт только идентификатор.
      properties:
        actor:
          type: string
        at:
          type: string
          format: date-time
        body:
          type: string
        clientId:
          type: string
          format: uuid
        dealId:
          type: string
          format: uuid
        direction:
          type: string
        id:
          type: string
          format: uuid
        kind:
          type: string
        leadId:
          type: string
          format: uuid
        subject:
          type: string
...
    AdminLeadConversion:
      type: object
      description: |
        Разбор заявки в сделку. Клиент либо указывается существующий,
        либо заводится из данных заявки.
      properties:
        amount:
          type: number
          minimum: 0
        clientId:
          type:
          - string
          - "null"
          format: uuid
          description: "Существующий клиент. Пусто — карточка клиента заводится из\
            \ данных заявки. Портал не ищет совпадения сам: слить две карточки пот\
            ом можно, разделить ошибочно слитые — уже нет."
        owner:
          type: string
          maxLength: 200
          minLength: 0
        pipeline:
          type: string
          enum:
          - sales
          - dealer
          - service
          example: sales
          minLength: 1
        title:
          type:
          - string
          - "null"
          description: Название сделки. Пусто — берётся из формы и изделия заявки.
          maxLength: 300
          minLength: 0
      required:
      - pipeline
...
    AdminNewDeal:
      type: object
      description: Новая сделка по уже заведённому клиенту.
      properties:
        amount:
          type: number
          example: 2650000.0
          minimum: 0
        clientId:
          type: string
          format: uuid
          description: "Клиент, по которому заводится сделка."
        currency:
          type: string
          description: Трёхбуквенный код валюты. Пусто — RUB.
          example: RUB
          pattern: "^$|^[A-Z]{3}$"
        owner:
          type: string
          maxLength: 200
          minLength: 0
        pipeline:
          type: string
          enum:
          - sales
          - dealer
          - service
          example: sales
          minLength: 1
        productSlug:
          type: string
        title:
          type: string
          example: Поставка двух систем VEDAL R2
          maxLength: 300
          minLength: 0
      required:
      - clientId
      - pipeline
      - title
    AdminNewInteraction:
      type: object
      description: Новая запись истории.
      properties:
        at:
          type:
          - string
          - "null"
          format: date-time
          description: "Когда это было. Пусто — сейчас. Звонок записывают после ра\
            зговора, и время разговора важнее времени записи."
        body:
          type: string
          description: "Что произошло. Это персональные данные: в топики и в журна\
            л текст не уходит."
          example: "Обсудили сроки: готовы отгрузить в марте, нужен аванс 30%."
          minLength: 1
        direction:
          type:
          - string
          - "null"
          description: "Направление: `in` — от клиента, `out` — клиенту. У заметки\
            \ направления нет."
          enum:
          - in
          - out
          pattern: ^$|in|out
        kind:
          type: string
          enum:
          - call
          - email
          - meeting
          - note
          example: call
          minLength: 1
          pattern: call|email|meeting|note
        subject:
          type: string
          example: Созвон по срокам поставки
          maxLength: 300
          minLength: 0
      required:
      - body
      - kind
...
    AdminQuoteRow:
      type: object
      description: Строка списка КП.
      properties:
        createdAt:
          type: string
          format: date-time
        currency:
          type: string
        dealId:
          type: string
          format: uuid
        dealTitle:
          type: string
        id:
          type: string
          format: uuid
        number:
          type: string
        sentAt:
          type: string
          format: date-time
        status:
          type: string
        total:
          type: number
        validUntil:
          type: string
          format: date
...
    Page:
      type: object
      description: Страница списка.
      properties:
        items:
          type: array
          description: Строки страницы.
          items:
            $ref: "#/components/schemas/AdminQuoteRow"
        page:
          type: integer
          format: int32
          description: "Номер страницы, с нуля."
          example: 0
        pages:
          type: integer
          format: int32
          description: Всего страниц.
          example: 3
        size:
          type: integer
          format: int32
          description: Размер страницы.
          example: 50
        total:
          type: integer
          format: int64
          description: Всего строк во всей выборке.
          example: 137
    PipelineView:
      type: object
      properties:
        lostStages:
          type: array
          description: "Стадии, которыми она заканчивается отказом. Чем воронка ко\
            нчается — правило домена, и форма берёт его отсюда: список, переписанн\
            ый в интерфейс, разъезжается с доменом на первой же новой воронке."
          items:
            type: string
        pipeline:
          type: string
        stages:
          type: array
          items:
            type: string
        wonStages:
          type: array
          description: "Стадии, которыми эта воронка заканчивается успехом."
          items:
            type: string
...
  securitySchemes:
    keycloak:
      bearerFormat: JWT
      description: |
        Токен доступа из Keycloak. Роли портала — `portal-admin`,
        `portal-sales` и `portal-production` в `realm_access.roles`; токен без них
        проходит проверку подписи и получает `403`.

        В режиме `vedal.iam.mode=local` вместо этого работает
        HTTP Basic поверх учётных записей портала — режим
        разработки, не контракт.
      scheme: bearer
      type: http